more notes. more reoganization. more fixes
