+++ data/config/settings.json
@@ -7,13 +7,13 @@
 		"data.scripts.starfarer.api.impl.campaign.rulemd"
 	],
 	#logs active (for debuging)
-	"AIRetrofit_logs": false,
+	"AIRetrofit_logs": true,
 
 	#option to startwith/always have new hullmods and abilitys from the start.
 	"AIRetrofit_alwaysGiveSkillsAndHullmods": false,
 
 	#this should be disabled for now. it may cause issues with invasions launched from AI-Pop worlds in nexerlin
-	"AIRetrofits_SwapAICrew": false,
+	"AIRetrofits_SwapAICrew": true,
 
 	#settings for AIRetrofits hullmod
 	"AIRetrofits_AIretrofit_SUPPLY_USE_MULT":1.0,#and by multi, i mean increase multi. zero whould be no supply change. -1 would be no supply cost.
