+++ data/strings/descriptions.csv
@@ -7,3 +7,11 @@ AIretrofit_maintainsPacts,RESOURCE,"Various parts for robot maintenance, for whe
 AIretrofit_SubCommandNode,RESOURCE,"An AI Command Core. While it is not even comparable to something like a gamma core, it's still capable of performing somewhat complicated tasks, provided there is only a small amount of factors to consider. If given too much to do, the core will lag to the point of uselessness.",,,,
 AIretrofit_roboticReplacementParts,RESOURCE,A set of easily-adaptable parts to repair robotic workers that have suffered extensive damage.,,,,
 AIretrofit_humanInterfaceNode,RESOURCE,"While you can't tell by looking at it, this 'human' is not, in fact a human. It's a robotic worker disguised as one. If no colonist can converse with a visitor, then this will naturally rouse suspicion. This 'person' helps to prevent that by talking to people; about their day, about work, about all sorts of things...",,,,
+,,,,,,
+AIretrofit_Advanced_WorkerDrone,RESOURCE,"temp: smaller, and stronger. 0.75, 1.25",,,,
+AIretrofit_Advanced_CombatDrone,RESOURCE,"temp: larger. stronger. 0.75, 1.25",,,,
+AIretrofit_Advanced_SurveyDrone,RESOURCE,"temp: smaller. same. 0.3, whatever the old one had i forgot",,,,
+,,,,,,
+AIretrofit_Omega_WorkerDrone,RESOURCE,"temp: smaler. stronger. 0.5, 3",,,,
+AIretrofit_Omega_CombatDrone,RESOURCE,"temp: larger. stronger. 1, 5",,,,
+AIretrofit_Omega_SurveyDrone,RESOURCE,"temp: smaler. stronger. 0.55, 4.5",,,,
