--- jars/src/data/scripts/AIWorldCode/industries/AIRetrofit_basicRobotManufactory.java
@@ -1,11 +1,0 @@
-package data.scripts.AIWorldCode.industries;
-
-import data.scripts.AIWorldCode.industries.personalRobotForge.AIRetrofit_PersonalRobotManufactoryBase;
-
-public class AIRetrofit_basicRobotManufactory extends AIRetrofit_PersonalRobotManufactoryBase {
-    @Override
-    public boolean isAvailableToBuild(){
-        return false;
-    }
-
-}
