+++ jars/src/desing/playerFleet/robots
@@ -1,0 +1,42 @@
+robots:
+    3 types:
+        salvage
+            -found markets.
+            -salvage
+            -repair conal hypershunt
+        combat
+            -raid
+            -invasions
+            -steal a starship
+        survey
+            -survey worlds.
+
+
+    3 tiers of each:(size,power)
+        salvage:
+            T1: 1    ,1
+            T2: 0.75 ,1.25
+            T3: 0.5  ,3
+        combat:
+            T1: 0.5  ,0.5
+            T2: 0.75 ,1.25
+            T3: 1,   5
+        survey:
+            T1: 1.1  ,1.5
+            T2: 0.3  ,1.5
+            T3: 0.55 ,4.5
+
+    modifiers:
+        power:
+            all robots are augmented by the strongest AI core in you fleet, and are penalized when there is none.
+            (all modifiers are a multiplayer on the normal amount of power per robot.)
+            AI-Core stats:
+                none:               0.85
+                sub-command-node:   1
+                gamma:              1.15
+                beta:               1.3
+                alpha:              1.45
+                omega:              1.6
+        defence:
+            currently, defence is the same as power, however i might remove or reduce the defence bonus,
+            so you lose and have to replace robots more often.
